next commit

+++ src/api/notes/handler.js
@@ -16,8 +16,11 @@ class NoteHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = undefined, body, tags } = request.payload;
+      const { id: cridentialId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: cridentialId,
+      });
 
       const response = h.response({
         status: 'success',
@@ -46,8 +49,9 @@ class NoteHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: cridentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(cridentialId);
     return {
       status: 'success',
       data: {
@@ -59,9 +63,9 @@ class NoteHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
-
+      const { id: cridentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, cridentialId);
       const note = await this._service.getNoteById(id);
-
       return {
         status: 'success',
         data: {
@@ -91,6 +95,8 @@ class NoteHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: cridentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, cridentialId);
       await this._service.editNoteById(id, request.payload);
       return {
         status: 'success',
@@ -119,6 +125,8 @@ class NoteHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: cridentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, cridentialId);
       await this._service.deleteNoteById(id);
       return {
         status: 'success',
